Password resets are working now
Lots of help from this tutorial:
- https://learndjango.com/tutorials/django-password-reset-tutorial

Removed the 'app_name' namespace because we would have to subclass
all of the generic templates being used by django to account for
our new logins.

We could move all the auth urls into the top-level 'portal/urls.py',
but then we would have to add some config to look for all the templates
in another place.