Add form page for editing user information
Pretty loosey-goosey over here but the good news is that selected
user info is editable. Next up, we want to do some permission
wrangling stuff so that a user can only edit their own info, but
that can wait for later.

Editing a user returns us to the user profile now.