first pass
Adds 2 modules, requests and dotenv

We use dotenv to load values from a .env file that currently has place holders. After this commit I'll add it to .gitignore so that we can modify our local versions without having to worry about ignoring it all the time.

The view is pretty simple, get the code from the remote server, add a space in the middle and use the code.html template (which got moved under /code/ for an arbitrary reason.