Gracefully handle missing auth token
Without an auth token, the /code route would fail every time with
an error.

Added an `if` check so that you can still boot up and develop the
app without a bearer token.

In dev mode, the app should be able to run by itself, without
a downstream dependency on another service that we don't own.