Override default AuthenticationForm for our login
The default in Django for some reason is to prefix all labels with
colons, so subclassing the form lets us remove those.
We can also add classes or whatever to the labels if we need, but
it's okay right now.

Found a nice Stackoverflow about subclassing and a Medium post
about removing colons.

Source:
- https://stackoverflow.com/questions/43542442/django-adding-placeholders-to-django-built-in-login-forms
- https://medium.com/@MatheusCAS/overriding-globally-field-label-suffix-in-django-ed3ffb10af7