Add expiry date to code page in en/fr
Added an "expires" string to the '/code' page with the current
time, in both English and French formatting.

The reason I'm handling the time formatting in the template is that
modifying the timezone on the server caused the time localization to
stop working.
In addition, the template `date` filter options are are more
numerous and flexible.

Using EST time for this PR, changing timezones is outside the scope
of the story.
The short version is that to get a user's timezone we either have
to use JS to guess the locale, or get the user to input their timezone
in some kind of settings menu.

Generally the less we have to deal with this stuff, the better.

Sources:
- https://docs.djangoproject.com/en/3.0/topics/i18n/timezones/
- https://tommikaikkonen.github.io/timezones/
- https://docs.djangoproject.com/en/3.0/ref/templates/builtins/#date
- https://stackoverflow.com/questions/10801397/system-date-formatting-not-using-django-locale