Don't double red line forms
When we have a "form error" (like "email and password do not belong
to a user"), we give the whole form a red border.
Whereas when we have an error with one specific field ("email must
be less than 200 words"), we give that field a red border.

When we had both, the borders would stack. Now they don't.