Stack header nav links on mobile
Our site looks pretty good on mobile except for the top nav,
with a broken layout that's one of the first things you see.

This CSS isn't meant to be the final design or anything, but
it improves our top nav layout on mobile until such time as
we want to solve this some other way.