Minor CSS fixes to form validation + nav links in header (#53)
* Don't double red line forms

When we have a "form error" (like "email and password do not belong
to a user"), we give the whole form a red border.
Whereas when we have an error with one specific field ("email must
be less than 200 words"), we give that field a red border.

When we had both, the borders would stack. Now they don't.

* Stack header nav links on mobile

Our site looks pretty good on mobile except for the top nav,
with a broken layout that's one of the first things you see.

This CSS isn't meant to be the final design or anything, but
it improves our top nav layout on mobile until such time as
we want to solve this some other way.