Add in the legit images
Found a string appending trick in a stackoverflow answer allowing us
to swap out part of the static filename depending on the language code.

Think it looks pretty swank, so I also used it for the FIP header.

Added empty alt tags because the images are primarily a visual aid
in order to be able to understand what the other patient is seeing

Adding image descriptions that would say "screen 1" or "screen 2" or
something feels like it's not that important.

Source:
- https://stackoverflow.com/a/49526265