Add a form to the start page to get to the key page
This is in response to a security story where we want to restrict
access to the "/key" page so that logged-in browsers can't be linked
to that page.

If someone tries to visit the "key" page using a GET request, they
will be redirected.

If they visit with a malformed POST request, they will see a CSRF
error page.

Added some tests and updated the ones that were breaking.

The request context isn't included on the CSRF error page for some
reason, but I don't think it's a dealbreaker.`

Sources:
- https://stackoverflow.com/questions/26925244/django-how-to-override-the-csrf-failure-template
- https://git.hci.cs.jmu.edu/lablablab/edu-git/edugit-django/blob/fc9f7b747403909b9ca2788096dfe60342439ce3/edugit/templates/403_csrf.html