Add an app version: 1.0.0
Added a VERSION file with the version number for the app.

Added the django-project-version library and set the DJVERSION_VERSION
settings variable to be the same as what we have in the VERSION file.

- Replace "Ok" on the /status page with the app version number
- Add the app version number in the <head> of the HTML

Finally, added a CHANGELOG file, starting at this most recent change.

Note on starting at V1: I asked @sidewalkballet about what version to
use (should we wait until "ATO" for V1) but she said not to worry about it
the ATO doesn't change anything about the product that users will notice,
so we shouldn't peg our versioning around it.

Since we're live now and have users, we can just say that now we're
on version 1.

Sources:
- https://pypi.org/project/django-project-version/