Add skiplink for logged in users
It's an a11y best practice and we should just put it in.

If you're not logged in, there's no menu to speak of, so we don't need
to "skip to the content".

Doesn't target the "main" element itself, because we want to skip the
"logged-in menu" if it exists.

Lifted it directly from what @brdunfield did on a past PR for CRA.

https://github.com/cds-snc/cra-claim-tax-benefits/pull/410