Add aria-describedby attribute if validation errors exist
"aria-describedby" attributes tell screen readers to read more
contextual information about a given field.

If someone is using a screenreader and they are pressing "tab" to get
to different form fields, they should be made aware when they reach
a field with a validation error.

This way, the validation error will be read after the form label.

Why not use aria-invalid or others?

The short answer is that I followed what the GDS design system is doing.
- eg: https://design-system.service.gov.uk/components/error-message/#label

You can read more here if you like:
- https://developer.paciellogroup.com/blog/2018/09/describing-aria-describedby/