(Super)admins no longer able to chagne others' passwords
This was something we secretly allowed, but are now removing as there's
no viable use case for it.

Also this removes the "yubikey" row that superadmins would see on
every other user's profile page (it was always false).

Only superadmins can see the "Yubikey" row, and only on their own account page.