Remove "required" attribute from input fields
Rqequired fields will still be validated on the serverside, but
removing the "required" attribute from from inputs stops the browser
from applying its own client-side validation.

The client-side validation applied by the browser is easy to get
around and doesn't change languages with the rest of the site, so
it's better to just remove it.

Source:
- https://docs.djangoproject.com/en/3.1/ref/forms/api/#django.forms.Form.use_required_attribute