Move backup codes logic into its own Django app
Created a "backup_code" app for the backup codes themselves.
The views.py and tests.py files are getting pretty long, so
let's move the backup code logic out of the "profiles" app.

It's worth noting there are lots of different candidates for
what we call these things, for example:

- Static codes
- Security codes
- Offline codes
- Backup codes
- Recovery codes
- Emergency codes
- One-time codes
- "Scratch cards"

But we're going with Backup codes here because the communicate
the intent of the codes (rather than "static codes" or "security codes").

A final decision on these won't be made until @YedidaZalik is back
from her time off, so it's possible this will change again in future.
(Which is fine, we can deal.)