Remove spaces and lowercase submitted 2FA codes
We display codes in uppercase and with a space in the middle,
so that they are easier to copy and remember is the idea.

However, we only accept lowercased codes without spaces in between them.

This PR introduces a method that strips whitespace and lowercases all
values submitted through the 2FA form.