Extend the phonetic templatetag
We want to display the phonetic alphabet for admins to speak codes
over the phone to users, but the alphabet was missing a few letters
and included an intro line that was specifically about patients.

Added the opportunity to submit a generic intro line.
Also added the missing letters as well because the backup codes can
contain all characters in the alphabet.

Source:
- https://docs.google.com/document/d/1vcmlz-2xabi94T-eLxXtBJR9HexeXGR0Gg0AAOjbB1c/edit