Add print styles for backup_codes pages
Added print styles only to the pages that a user would see when
getting a backup code for themselves (not going to add it to the
admin one).
For all other pages of the site, printing will keep all the visible
elements, including the header and footers.

For the backup codes pages, the print styles:
- hide the header, footer, Canada FIP elements, the on-page buttons, and the navs
- show a COVID alert logo image
- show an alternative title with the email address of the logged-in user
- bump up the font sizes for everything.

Didn't use different text than what was on the screen, as that would
have been a lot more logic than I wanted to do.