Remove SMTP variables, set default EMAIL_BACKEND to "console"
Since we are using Notify for all our emails (that we know of), we
have moved away from SMTP.

It's possible that some current or future edge case still uses the
default Django send_mail function so let's use the console output
for that should it happen.

Removed all of the environment variables related to the email backend,
including removing it from the ENV configuration.

Also updated the README to remove references to SMTP, and reordered
the env variables so that Notify is up a lot higher.