add test + bad code
It's TDD this time.

Wrote a test to see if we can delete a user once they have a backup
code, and it calls the exception.
Then I wrote some code to try and fix the behaviour and it doesn't work.

Now I'm out of time, so I'm going to park it for now.

Doesn't look like the code I wrote is helpful, but this SO answer might be it:
- https://stackoverflow.com/a/49772577