Reference SCSS from announcements module in the portal module
As I started building the new pages, I noticed what a faff it is
to have all the CSS in the `/profiles` subfolder, but it's also
a faff to have to compile the announcements module CSS separately
from the rest of the app.

Went into the `django-sass` package to see what their recommendation
and they seem to be recommending that you link your dependencies
between modules:

```
@import 'app1/scss/colors';
@import 'app2/scss/colors';
// Or since you are in app2, you can reference its colors with a relative path.
@import 'colors';
```

Basically, we have a few options here but I think that it's fair
to make announcements less independent if it makes our code more
comprehensible (and solves the problem of adding CSS in other modules).