Split start page into 2 screens
Part of the reasoning why people occasionally create keys without
meaning to is that it's just easy to click that button unintentionally.

The proposed solution here is to create 2 pages.
1. is this patient positive and have the app on their phone?
2. ready to enter the code or write it down somewhere

And then after that we show the key.

Also changing the "start" url in the man nav, although we're
keeping the URL path and the internal "start" references.