CSS updates for new start pages
Added a number of different CSS fixes in this commit.

1. New chevrons

Added a "go back" link with a leftwards-pointing chevron, and
a "start" button with a rightwards pointing chevron.

Made the chevron CSS into a mixin that we can attach to other
elements, as it's not really a standalone component.

2. "start" button styles

Like GOV.UK, we are adopting a green "start" button. Mostly
copied the GDS design system button colours here and added
a chevron.

3. .content-section

This is kind of shit but basically, we need some layout helpers.
Used the "content-button" precendent here. The point of this class
is that we can separate things on the page like in the Figmas.

4. Lists

Did some CSS stuff to make lists a but cleaner.
Basically, we have regular list CSS, and then "expanded" list CSS
for when lists have a bunch of sentences rather than a few words each.

Very subtle changes but the "admin accounts" about page and the
"start" page look better now.

Sources:
- GDS button: https://design-system.service.gov.uk/components/button/