Create 2 lists for backup codes
Instead of creating one list for backup codes, where each list item
has a visually hidden code and a visually visible code, let's have 2 lists.

One list is hidden, and one list is visible.

The problem with the last approach was that selecting all of the codes at once would duplicate the codes. All the visually-hidden
codes were still being selected.

The user-select CSS doesn't let you select the text directly, but
if you selected the text before and after it, the intermediate text is selected regardless.
You can see that here: https://codepen.io/chriscoyier/pen/ZGKYxj

Tried a bunch of CSS here and a bunch of JS and it was just easier
to have 2 lists and have that be the end of it.

No visual changes for users, but copying and pasting will work better.