Add a confirmation page
The start page and the confirmation page are the two static pages
here, so I've added them both first.

We used to have an ".info-message" component on the "start" page
of the app but we don't anymore. Used it for the confirmation page, and
rewrote some of the CSS to be more like the figma mocks.